Change Matrix to use float64 for allow more precision for geocentric coordinates

+++ const.go
@@ -4,7 +4,7 @@ import "encoding/json"
 
 var (
 	// DefaultMatrix defines an identity matrix.
-	DefaultMatrix = [16]float32{1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1}
+	DefaultMatrix = [16]float64{1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1}
 	// DefaultRotation defines a quaternion without rotation.
 	DefaultRotation = [4]float32{0, 0, 0, 1}
 	// DefaultScale defines a scaling that does not modify the size of the object.
@@ -26,7 +26,7 @@ const (
 )
 
 var (
-	emptyMatrix   = [16]float32{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}
+	emptyMatrix   = [16]float64{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}
 	emptyRotation = [4]float32{0, 0, 0, 0}
 	emptyScale    = [3]float32{0, 0, 0}
 )
